Landing Page Mobile Version Fixes #5
